Track connections to upstream servers
This isn't a perfect solution. The upstream vts node is only created
once a request has finished. This means if this upstream only receives
long-lasting websocket connections, there'll be a period of time where
we report 0 connections when there's actually a non-0 amount.

For our use-case this is fine, as the server also exposes some metrics
endpoints which are periodically scraped. However, for something like
copy-cat this may be more problematic.